Update README.md
This is my first public Code Submission and there are some bugs and features I Need to address.
Input and contributions are welcome.

Simply drag the Calendar prefab into the hierarchy.

This calendar is set to use leap years by default from Roman Times(709AD) and well into the future.
It uses both the Julian and Gregorian leap year systems  and can switch between them automatically depending on the current in game year.

How to use:

Drag and drop the GameCalendar prefab into the hierarchy and click on it.

In the Start Year Field enter a year, next choose the starting month, day of month(like the 14th or whatever) and the day of the week.
Test it out by hitting play in the editor.

Currently I have only made functions to change the current day by +1. 
you can test this by pressing 1 while in play mode to change the current day by +1.

The Calendar will update automatically.

The Seasons are calculated automatically depending on the current month.

I will Continue to develop and refactor this Calendar system over time.